r7rs-lang-plan R10: chibi's R7RS suite as a pass count, 1036/1216 on both back ends
tests/r7rs/run-conformance.py runs chibi-scheme's r7rs-tests.scm (vendored
with its BSD licence) as one program per round, splitting it into marked
top-level forms under a (chibi test)-compatible harness: a rejected form is
dropped by its diagnostic line, a form that stops the program is counted
failed and the run resumes after it. tests/run-r7rs-conformance.sh and the
ctest target tur_r7rs_conformance run both back ends with a floor of 1036.

The first run passed 887 on the interpreter and the compiled program did
not build. Fixed on the way, each pinned in tests/fixtures:

- assignment conversion in scheme_lower.c: a set! variable a lambda captures
  lives in an R7rsBox, so a do loop or for-each lambda that set!s an outer
  variable is right when compiled (it answered 0);
- user binders named like Turmeric special forms (return) and top-level
  defines of auto-loaded stdlib names (list-length) are renamed;
- vector literals are constants; quasiquote fires an unquote under a quote
  and reads the long forms; syntax-rules literal `_` and literal ellipsis;
  internal defines are letrec*; letrec lambdas that take each other as
  values become cells; `(define first car)` and operators as values are any;
- inexact integer division, inexact numerator/denominator, case-insensitive
  inf/nan and R5RS exponent markers, catchable apply errors, symbol bars,
  make-bytevector's optional fill, `\"` in |symbols|;
- outside the dialect: a record-typed parameter forward-declared as int; the
  interpreter's lifted lambdas not seeing an enclosing letrec; a global def
  holding a closure read in C before its declaration; +inf.0 emitted as
  `inf`; the closure-env registry's uint8_t counter crashing tur at 256
  envs; fd_for_binding's quadratic scan (emit-c of the suite 4.5 min -> 42 s).

r7rs-control's expected output had recorded the `return` bug (#f for 4) and
is corrected. The typed/Saffron half of the capture divergence is filed as
docs/reported/compiled-closure-copies-a-captured-mut.md.

+++ src/compiler/emit_expr.c
@@ -53,7 +53,7 @@ void emit_env_struct_register(EmitCtx *ctx, const Symbol *env_name,
  * -- the registry owns it. */
 const char *emit_env_struct_fn_typedef(EmitCtx *ctx, const Symbol *env_name) {
     if (!ctx->env_struct_names || !ctx->env_struct_fn_typedefs) return NULL;
-    for (uint8_t i = 0; i < ctx->n_env_struct_names; i++)
+    for (uint32_t i = 0; i < ctx->n_env_struct_names; i++)
         if (ctx->env_struct_names[i] == env_name)
             return ctx->env_struct_fn_typedefs[i];
     return NULL;
@@ -65,7 +65,7 @@ void emit_closure_env_struct_and_glue(EmitCtx *ctx, Buf *out,
                                       bool resolve_spec_params) {
     bool already_emitted = false;
     if (ctx->env_struct_names) {
-        for (uint8_t i = 0; i < ctx->n_env_struct_names; i++) {
+        for (uint32_t i = 0; i < ctx->n_env_struct_names; i++) {
             if (ctx->env_struct_names[i] == env_name) {
                 already_emitted = true;
                 break;
